fix potential SIGSEGV while removing entry from hash table: this concerns the rozofs_cachetack and the rozofs_kpi that were using as a key a pointer to a FID that was stored in the ientry.
The issue may happen where the ientry is released. In that case the
FID pointer might point on a memory array that is either unaccessible or
contains a value that is not the initial FID. In the last case, this
prevent the hash entry removing, that leads to a potential SIGSEGV
on the next LRU update of the entry.