Fix for the issue where storcli takes 100% of a core
The issue comes from the fact that storcli cannot handle the reception
of a message coming from rozofsmount. Since the socket has a priority of
16, we do not re-evaluate if the socket can be polled or not.
By setting the priority to 3, we take care of that case and the socket is
removed from the active socket until we get enough ressources to go on.
That happens only when there are 4 storclis configured.