rozofs_write: Fix a bug that could trigger a kernel panic because of a NULL pointer.
The issue was coming from the fact that the "padding" field was tested without taking care of the bitfields it contains.
As a consequence, it might be possible that a bit is asserted in the fuse request in the cache of a flush of the writeback caceh for a page.

In that case the pointer to the fuse write request in not saved, so the client
on the userspace side thinks that the write whould be done by using fuse IOCTL 5
but the array that points to the fuse request is NULL. As a consequence, at the time
rozofsmount attempt to get the data to write thanks the IOCTL, it gets a SIGSEGV in the kernel.