rozofsmount:
Fix an issue on the size of the rozofsmount/storcli shared buffer.
4KB have been added to the buffer size. It is necessary to avoid
bumping into the next buffer when RozoFS is configured in direct I/O mode.

In direct I/O, if the user buffer is not aligned on a 4KB boundary, we might
need to read the max size (512KB or 1MB) plus 4KB more.